+++ include/wave/geometry/src/geometry/leaf/CompactRigidTransform.hpp
@@ -13,26 +13,19 @@ namespace wave {
  * The first four coefficients are a quaternion, stored in Eigen::Quaternion order. The
  * last three coefficients are a translation. The memory is contiguous.
  *
- * @tparam ImplType The type to use for storage (e.g. Eigen::Matrix<double, 7, 1>)
- *
+ * @tparam QuatType The underlying type of quaternion (e.g. Eigen::Quaterniond)
+ * @tparam VecType The underlying type of translation (e.g. Eigen::Vector3d)
  * The alias RigidTransformQd is provided for the typical scalar type, double.
  */
-template <typename ImplType>
+template <typename QuatType, typename VecType>
 class CompactRigidTransform
-    : public RigidTransformBase<CompactRigidTransform<ImplType>>,
-      public LeafStorage<CompactRigidTransform<ImplType>, ImplType> {
-    static_assert(internal::is_eigen_vector<7, ImplType>::value,
-                  "ImplType must be an Eigen 7-vector type.");
-
-    using Storage = LeafStorage<CompactRigidTransform<ImplType>, ImplType>;
-    using Scalar = typename Eigen::internal::traits<ImplType>::Scalar;
-
-    // Eigen Blocks for nested rotation and translation
-    using RotationMap = Eigen::Map<Eigen::Quaternion<Scalar>>;
-    using RotationConstMap = Eigen::Map<const Eigen::Quaternion<Scalar>>;
-    using TranslationBlock = Eigen::Block<ImplType, 3, 1>;
-    using TranslationConstBlock = Eigen::Block<const ImplType, 3, 1>;
-
+    : public RigidTransformBase<CompactRigidTransform<QuatType, VecType>>,
+      public CompoundLeafStorage<CompactRigidTransform<QuatType, VecType>,
+                                 QuaternionRotation<QuatType>,
+                                 Translation<VecType>> {
+    using Storage = CompoundLeafStorage<CompactRigidTransform<QuatType, VecType>,
+                                        QuaternionRotation<QuatType>,
+                                        Translation<VecType>>;
 
  public:
     // Inherit constructors from LeafStorage
@@ -50,78 +43,82 @@ class CompactRigidTransform
                           const TranslationBase<TDerived> &t) {
         this->rotationBlock() = R.derived();
         this->translationBlock() = t.derived();
-    };
+    }
 
     /** Constructs from a rotation and translation, given as Eigen matrices */
     template <typename RDerived, typename TDerived>
     CompactRigidTransform(const Eigen::MatrixBase<RDerived> &R,
-                          const Eigen::MatrixBase<TDerived> &t) {
+                          const Eigen::MatrixBase<TDerived> &t)
+        : Storage{QuatType{R.derived()}, t.derived()} {
         EIGEN_STATIC_ASSERT_MATRIX_SPECIFIC_SIZE(RDerived, 3, 3);
         EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(TDerived, 3);
-        this->value() << Eigen::Quaternion<Scalar>{R}.coeffs(), t;
-    };
+    }
 
     /** Construct from an Eigen rotation and translation vector */
     template <typename RDerived, typename TDerived>
     CompactRigidTransform(const Eigen::RotationBase<RDerived, 3> &q,
                           const Eigen::MatrixBase<TDerived> &t)
-        : CompactRigidTransform{q.matrix(), t} {};
-
-    /** Construct from an Eigen 7-vector */
-    template <typename VDerived, TICK_REQUIRES(internal::is_eigen_vector<7, VDerived>{})>
-    explicit CompactRigidTransform(const Eigen::MatrixBase<VDerived> &v)
-        : Storage{typename Storage::init_storage{}, v.derived()} {
-        EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(VDerived, 7);
-    }
+        : CompactRigidTransform{q, t} {
+        EIGEN_STATIC_ASSERT_VECTOR_SPECIFIC_SIZE(TDerived, 3);
+    };
 
     // block getters - for differentiable expressions and consistent interface, use
     // .rotation() and .translation()
     // @todo make private?
 
     /** Returns a reference to the rotation portion of this transform */
-    QuaternionRotation<RotationMap> rotationBlock() noexcept {
-        return QuaternionRotation<RotationMap>{RotationMap{this->value().data()}};
+    auto &rotationBlock() noexcept {
+        return std::get<0>(this->value());
     }
 
     /** Returns a const reference to the rotation portion of this transform */
-    QuaternionRotation<RotationConstMap> rotationBlock() const noexcept {
-        return QuaternionRotation<RotationConstMap>{
-          RotationConstMap{this->value().data()}};
+    const auto &rotationBlock() const noexcept {
+        return std::get<0>(this->value());
     }
 
     /** Returns a reference to the translation portion of this transform */
-    Translation<TranslationBlock> translationBlock() noexcept {
-        return Translation<TranslationBlock>{this->value().template tail<3>()};
+    auto &translationBlock() noexcept {
+        return std::get<1>(this->value());
     }
 
     /** Returns a const reference to the translation portion of this transform */
-    Translation<TranslationConstBlock> translationBlock() const noexcept {
-        return Translation<TranslationConstBlock>{this->value().template tail<3>()};
+    const auto &translationBlock() const noexcept {
+        return std::get<1>(this->value());
     }
 };
 
 namespace internal {
 
-template <typename ImplType>
-struct traits<CompactRigidTransform<ImplType>>
-    : rt_leaf_traits_base<CompactRigidTransform<ImplType>> {
-    using PlainType = CompactRigidTransform<typename ImplType::PlainObject>;
+template <typename Q, typename V>
+struct traits<CompactRigidTransform<Q, V>>
+    : compound_leaf_traits_base<CompactRigidTransform<Q, V>,
+                                QuaternionRotation<Q>,
+                                Translation<V>>,
+      frameable_transform_traits {
+ private:
+    using TraitsBase = compound_leaf_traits_base<CompactRigidTransform<Q, V>,
+                                                 QuaternionRotation<Q>,
+                                                 Translation<V>>;
+
+ public:
+    using TangentType = Twist<Eigen::Matrix<typename TraitsBase::Scalar, 6, 1>>;
 };
 
 /** Converts from compact to matrix rigid transform
  */
-template <typename ToImpl, typename FromImpl>
+template <typename ToImpl, typename QuatType, typename VecType>
 auto evalImpl(expr<Convert, MatrixRigidTransform<ToImpl>>,
-              const CompactRigidTransform<FromImpl> &rhs) {
-    return MatrixRigidTransform<ToImpl>{rhs.rotation(), rhs.translation()};
+              const CompactRigidTransform<QuatType, VecType> &rhs) {
+    return MatrixRigidTransform<ToImpl>{rhs.rotationBlock(), rhs.translationBlock()};
 }
 
 /** Converts from rotation matrix to quaternion
  */
-template <typename ToImpl, typename FromImpl>
-auto evalImpl(expr<Convert, CompactRigidTransform<ToImpl>>,
+template <typename QuatType, typename VecType, typename FromImpl>
+auto evalImpl(expr<Convert, CompactRigidTransform<QuatType, VecType>>,
               const MatrixRigidTransform<FromImpl> &rhs) {
-    return CompactRigidTransform<ToImpl>{rhs.rotation(), rhs.translation()};
+    return CompactRigidTransform<QuatType, VecType>{rhs.rotationBlock(),
+                                                    rhs.translationBlock()};
 }
 
 /** Implements "conversion" between CompactRigidTransform types
@@ -129,10 +126,10 @@ auto evalImpl(expr<Convert, CompactRigidTransform<ToImpl>>,
  * While this seems trivial, it is needed for the case the template params are not the
  * same.
  */
-template <typename ToImpl, typename FromImpl>
-auto evalImpl(expr<Convert, CompactRigidTransform<ToImpl>>,
-              const CompactRigidTransform<FromImpl> &rhs) {
-    return CompactRigidTransform<ToImpl>{rhs.value()};
+template <typename ToQ, typename ToV, typename FromQ, typename FromV>
+auto evalImpl(expr<Convert, CompactRigidTransform<ToQ, ToV>>,
+              const CompactRigidTransform<FromQ, FromV> &rhs) {
+    return CompactRigidTransform<ToQ, ToV>{rhs.rotationBlock(), rhs.translationBlock()};
 }
 
 
@@ -140,7 +137,7 @@ auto evalImpl(expr<Convert, CompactRigidTransform<ToImpl>>,
 
 // Convenience typedefs
 
-using RigidTransformQd = CompactRigidTransform<Eigen::Matrix<double, 7, 1>>;
+using RigidTransformQd = CompactRigidTransform<Eigen::Quaterniond, Eigen::Vector3d>;
 
 template <typename F1, typename F2>
 using RigidTransformQFd = Framed<RigidTransformQd, F1, F2>;